+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Logging out now calls the logout route on the backend
+
 ## [1.0.0] - 2020-12-20
 
 ### Added
 
--   Initial logged release
+- Initial logged release